refactor(main): split main page into components and use chat as main page
